Move SASL check after STARTTLS
According to Dovecot documentation, this is not always available before
STARTTLS:

The SASL capability lists the available SASL authentication mechanisms.
If this list is empty and STARTTLS is available, it probably means that
the server forces you to initiate TLS first (as dictated by
auth_allow_cleartext = yes in dovecot.conf).